rename: 홈 컴포넌트 폴더 components 폴더로 이동
