+++ src/components/home/PortfolioRatioComponent.jsx
