+++ src/components/home/RobotAnalyzing.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import logo from "../../../assets/logo.svg";
+import logo from "../../assets/logo.svg";
 
 export default function RobotAnalyzing() {
   return (
